Move workspace structure logging to the Workspaces layer
